Remove tool classes and usage examples
